+++ modules/tinymce/CHANGELOG.md
@@ -4,10 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-(Post 6.3) - temporary
-
 ### Added
-- New `ui_of_tomorrow` setting for TBA
+- New `ui_mode` setting for editor in scrollable containers support. #TINY-9414
 
 ## Unreleased